// src/app/page/master/infeksi-nosokomial/infeksi-nosokomial.interface.ts
import {EventEmitter} from '@angular/core';

export interface InfeksiNosokomial {
	namaHari?: string;
	namaInfeksiNosokomial?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	id?: any;
	kdJenisInfeksiNosokomial?: number;
	version?: any;
	kode?: number;
}